Updates for Xcode 4.4 and Apple LLVM Compiler 4.0 (~= clang 3.1): * Fixed numerous format string problems, many of them real bugs. * Added format attribute to OOLog, and fixed more format string problems. * Cleaned up misuse of +[NSString stringWithString:]. * Updated OS X SDK to 10.7. Deployment target is still 10.5, although ideally   it should bet 10.6 for 64-bit. (See also long comment in   oolite-targets.xcconfig.)
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@5139 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Scripting/OOJSEngineDebuggerHelpers.m
@@ -146,7 +146,7 @@ const char *JSValueToStrSafeDbg(jsval val)
 {
 	NSString *formatted = nil;
 	
-	if (JSVAL_IS_INT(val))			formatted = [NSString stringWithFormat:@"%i", (long)JSVAL_TO_INT(val)];
+	if (JSVAL_IS_INT(val))			formatted = [NSString stringWithFormat:@"%i", JSVAL_TO_INT(val)];
 	else if (JSVAL_IS_DOUBLE(val))	formatted = [NSString stringWithFormat:@"%g", JSVAL_TO_DOUBLE(val)];
 	else if (JSVAL_IS_BOOLEAN(val))	formatted = (JSVAL_TO_BOOLEAN(val)) ? @"true" : @"false";
 	else if (JSVAL_IS_STRING(val))
@@ -189,7 +189,7 @@ const char *JSIDToStrSafeDbg(jsid anID)
 {
 	NSString *formatted = nil;
 	
-	if (JSID_IS_INT(anID))			formatted = [NSString stringWithFormat:@"%i", (long)JSID_TO_INT(anID)];
+	if (JSID_IS_INT(anID))			formatted = [NSString stringWithFormat:@"%i", JSID_TO_INT(anID)];
 	else if (JSID_IS_VOID(anID))	return "void";
 	else if (JSID_IS_EMPTY(anID))	return "empty";
 	else if (JSID_IS_ZERO(anID))	return "0";
